add-joke-api-and-fix-frontend

+++ client/assets/helpers.js
@@ -1,8 +1,6 @@
 let todoId = null
-let statusTodo = null
 
 function showLoginPage(){
-    $("#login-page1").show()
     $("#navbar").hide()
     $("#login-page").show()
     $("#register-page").hide()
@@ -30,36 +28,62 @@ function onSignIn(googleUser) {
 
     $.ajax({
         method: "POST",
-        url: "http://localhost:3001/googleLogin",
+        url: "http://localhost:3000/googleLogin",
         data: { googleToken }
     })
     .done(response => {
+        Swal.fire({
+            title: "Succes!",
+            text: "Login google succes!",
+            icon: "success",
+            button: "Ok!",
+        });
         localStorage.setItem('access_token', response.access_token)
         showMainPage()
-        console.log(response)
     })
     .fail(xhr => {
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Something went wrong!'
+        })
         console.log(xhr, "<<< error")
     })
 }
-function login(){
+function login(swal){
+    console.log(swal, "<<<< swal bosss")
     const email = $("#email-login").val()
     const password = $("#password-login").val()
     $.ajax({
         method: "POST",
-        url: "http://localhost:3001/login",
+        url: "http://localhost:3000/login",
         data: {
             email,
             password
         }
     })
     .done(response => {
+        Swal.fire({
+            title: "Good job!",
+            text: "You clicked the button!",
+            icon: "success",
+            button: "Aww yiss!",
+          });
         console.log(response)
         localStorage.setItem('access_token', response.access_token)
         showMainPage()
     })
     .fail((xhr, textStatus) => {
-        console.log(textStatus)
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Something went wrong!'
+        })
+        console.log(xhr.responseJSON.message)
+    })
+    .always(xhr => {
+        $("#email-login").val('')
+        $("#password-login").val('')
     })
 }
 function register(){
@@ -67,7 +91,7 @@ function register(){
     const password = $("#password-regis").val()
         $.ajax({
             method: "POST",
-            url: "http://localhost:3001/register",
+            url: "http://localhost:3000/register",
             data: {
                 email,
                 password
@@ -93,28 +117,28 @@ function fetchTodo(){
     $("#todo-list").empty()
     $.ajax({
         method: "GET",
-        url: "http://localhost:3001/todos",
+        url: "http://localhost:3000/todos",
         headers: {
             access_token: localStorage.getItem("access_token")
         }
     })
     .done(response => {
-        console.log(response)
         response.forEach(element => {
             $("#todo-list").append(`
-            <div class="col-3 mx-3">
-                <div class="card mt-5" style="width: 18rem;">
+            <div class="col-3">
+                <div class="card mt-5" style="width: 19rem;">
                 <div class="card-body">
-                <h5 class="card-title"><p class="text-muted">Title: </p>${element.title}</h5>
+                <h5 class="card-title">Title:  ${element.title}</h5>
                 </div>
                 <ul class="list-group list-group-flush">
-                <li class="list-group-item"><p class="text-muted">Date: </p>${element.due_date.substr(0, 10)}</li>
-                <li class="list-group-item"><p class="text-muted">Description: </p>${element.description}</li>
-                <li class="list-group-item"><p class="text-muted">status: </p>${element.status}</li>
+                <li class="list-group-item">Due Date:  ${element.due_date.substr(0, 10)}</li>
+                <li class="list-group-item">Description:  ${element.description}</li>
+                <li class="list-group-item">status:  ${element.status}</li>
                 </ul>
                 <div class="card-body">
-                <button class="btn btn-primary" onclick="editTodoForm(${element.id})" id="edit-todo"> Edit</button>
-                <button class="btn btn-danger" onclick="deleteTodo(${element.id})" id="delete-todo"> Delete</button> 
+                <button class="btn btn-outline-primary ms-3" onclick="editTodoForm(${element.id})" id="edit-todo"> Edit</button>
+                <button class="btn btn-outline-secondary ms-2" onclick="updateStatusForm(${element.id})" id="update-todo"> Update</button>
+                <button class="btn btn-outline-danger ms-2" onclick="deleteTodo(${element.id})" id="delete-todo"> Delete</button> 
                 </div>
                 </div>
             <div>
@@ -130,9 +154,10 @@ function addTodo(){
     const status = $("#status-add").val()
     statusTodo = status
 
+    
     $.ajax({
         method: "POST",
-        url: "http://localhost:3001/todos",
+        url: "http://localhost:3000/todos",
         headers: {
             access_token: localStorage.getItem("access_token")
         },
@@ -144,56 +169,171 @@ function addTodo(){
         }
     })
     .done(response => {
-
+        Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: `Todo ${response.title} Success Add`,
+            showConfirmButton: false,
+            timer: 1500
+          })
         fetchTodo()
         console.log(response)
     })
     .fail(xhr => {
+        console.log(xhr.responseJSON.message)
+        Swal.fire({
+            title: 'Error!',
+            text: 'Do you want to continue',
+            icon: 'error',
+            confirmButtonText: 'Cool'
+        })
+        $('#add-form').show()
         console.log(xhr)
     })
+    .always(_ => {
+        $("#title-add").val('')
+        $("#due_date-add").val('')
+        $("#add-description").val('')
+        $("#status-add").val('')
+    })
+    $.ajax({
+        method: "get",
+        url: 'http://localhost:3000/api/jooks',
+        headers:{
+            access_token: localStorage.getItem("access_token")
+        }
+    })
+        .done(response => {
+            console.log(response)
+            response.value.forEach(el => {
+                $('#title-modal').append(`
+                    Jooks
+                `)
+                $('.modal-body').append(`
+                    <h5>${el.joke}</h5>
+                `)
+            })
+        })
+        .fail(xhr => {
+            console.log(xhr)
+        })
 }
 function cancelEdit(event){
     event.preventDefault()
     console.log(event)
+    $("#new-task").show()
+    $("#title-listTodo").show()
     $("#todo-list").show()
-    $("#edit-form").hide()
+    $("#edit-form").empty()
 }
-function editTodoForm(id){
+function updateStatusForm (id) {
     todoId = id
     console.log(id)
     $.ajax({
+        method: 'get',
+        url: `http://localhost:3000/todos/${todoId}`,
+        headers: {
+            access_token: localStorage.getItem('access_token')
+        }
+    })
+        .done(response => {
+            $('#new-task').hide()
+            $('#title-listTodo').hide()
+            $('#todo-list').hide()
+            console.log(response)
+            $('#update-form').append(`
+                <h3 class="mt-5 mb-4 text-center">Update Status </h3>
+                <form id="update-status-todo">
+                    <div class="form group form-floating mt-2 mb-2" id="floating-status">
+                    <select id="status-update" class="form-select">
+                    <option value="unfinised" ${response.status === "unfinised" ? 'selected' : ''}>unfinised</option>
+                    <option value="finised" ${response.status === "finised" ? 'selected' : ''}>finished</option>
+                    </select>
+                    <label for="status-edit" class="">Status</label>
+                </div>
+                <button id="update-submit" class="btn btn-outline-primary mb-2" type="submit"> Submit</button>
+                </form>
+            
+            `)
+        })
+        .fail(xhr => {
+            console.log(xhr)
+        })
+}
+function updateStatus(id) {
+    const status = $('#status-update').val()
+    console.log(status, "<<<<<<<<")
+    todoId = id
+    console.log(todoId, '<<<< update status')
+    $.ajax({
+        method: 'patch',
+        url: `http://localhost:3000/todos/${todoId}`,
+        headers: {
+            access_token: localStorage.getItem('access_token')
+        },
+        data: {
+            status
+        }
+    })
+        .done(response => {
+            $("#new-task").show()
+            $("#title-listTodo").show()
+            $("#todo-list").show()
+            $("#update-form").empty()
+            Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: `Todo ${response.title} Success updated`,
+                showConfirmButton: false,
+                timer: 1500
+            })
+        })
+        .fail(xhr => {
+            console.log(xhr)
+        })
+}
+function editTodoForm(id){
+    todoId = id
+    console.log(todoId)
+    $.ajax({
         method: "GET",
-        url: `http://localhost:3001/todos/${todoId}`,
+        url: `http://localhost:3000/todos/${todoId}`,
         headers: {
             access_token: localStorage.getItem("access_token")
         }
     })
     .done(response => {
+        console.log(response.title)
+        $("#new-task").hide()
+        $("#title-listTodo").hide()
         $("#todo-list").hide()
         $("#edit-form").append(`
         <form id="formEdit mt-2">
             <div class="row">
+            <h3 class="mt-5 mb-4">Edit Todo</h3>
             <div class="col form-floating">
-            <input type="text" class="form-control title" placeholder="Title" id="titleEdit" value="${response.title}">
+            <input type="text" class="form-control title" id="titleEdit" value="${response.title}">
+            <label for="titleEdit" class="ms-3"> Title </label>
             </div>
-            <div class="col">
+            <div class="col form-floating">
             <input type="date" class="form-control due_date" placeholder="due date" id="dueDateEdit" value="${response.due_date.split('T')[0]}">
+            <label for="dueDateEdit" class="ms-3">Due-Date</label>
             </div>
-         </div>
-         <div class="form-group">
-         <div class="form-group w-100 mt-2">
-             <textarea class="form-control description" id="descriptionEdit" rows="3" placeholder="Description Here....">${response.description}</textarea>
-             </div>
-                <label > Status</label>
-                <select id="status-edit">
-                    <option checked></option>
-                    <option value="finised">finished</option>
-                    <option value="unfinised">unfinished</option>
-
+            <div class="form-group form-floating w-100 mt-2">
+                <textarea class="form-control description" id="descriptionEdit" rows="3" placeholder="Description Here....">${response.description}</textarea>
+                <label for="descriptionEdit" class="ms-3">Description</label>
+                </div>
+                <div class="form group form-floating mt-2 mb-2" id="floating-status">
+                <select id="status-edit" class="form-select">
+                <option value="unfinised" ${response.status === "unfinised" ? 'selected' : ''}>unfinised</option>
+                <option value="finised" ${response.status === "finised" ? 'selected' : ''}>finished</option>
                 </select>
-            </div>
-                <button type="submit" class="btn btn-primary" id="edit-todo">Edit</button>
-                <button id="cancel-edit" onclick="cancelEdit(event)" class="btn btn-danger">Cancel</button>
+                <label for="status-edit" class="ms-3">Status</label>
+
+               </div>
+         </div>
+                <button type="submit" class="btn btn-outline-primary ms-3" id="edit-todo">Edit</button>
+                <button id="cancel-edit" onclick="cancelEdit(event)" class="btn btn-outline-danger">Cancel</button>
             </form>
         `)  
         
@@ -208,10 +348,10 @@ function editTodo(){
     const due_date = $("#dueDateEdit").val()
     const description = $("#descriptionEdit").val()
     const status = $("#status-edit").val()
-    console.log(status, description, due_date, title, "<<<< edit todo")
+    console.log(status, description, due_date, title, "<<<< edit todo") 
     $.ajax({
         method: "PUT",
-        url: "http://localhost:3001/todos/" + todoId,
+        url: "http://localhost:3000/todos/" + todoId,
         headers: {
             access_token: localStorage.getItem("access_token")
         },
@@ -223,6 +363,10 @@ function editTodo(){
         }
     })
     .done(response => {
+        $("#new-task").show()
+        $("#title-listTodo").show()
+        $("#todo-list").show()
+        $("#edit-form").empty()
         fetchTodo()
         console.log(response)
     })
@@ -231,17 +375,71 @@ function editTodo(){
     })
 }
 function deleteTodo(id){
-    $.ajax({
-        method: "delete",
-        url: "http://localhost:3001" + "/todos/" + id,
-        headers: {
-            access_token: localStorage.access_token
+    const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+          confirmButton: 'btn btn-outline-primary ms-3',
+          cancelButton: 'btn btn-outline-danger'
+        },
+        buttonsStyling: false
+      })
+      
+      swalWithBootstrapButtons.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+            let todo= null
+           
+            $.ajax({
+                method: "get",
+                url: "http://localhost:3000" + "/todos/" + id,
+                headers: {
+                    access_token: localStorage.access_token
+                }
+            })
+                .done(response => {
+                    todo = response.title
+                })
+                .fail(xhr => {
+                    console.log(xhr)
+                })
+            $.ajax({
+                method: "delete",
+                url: "http://localhost:3000" + "/todos/" + id,
+                headers: {
+                    access_token: localStorage.access_token
+                }
+            })
+                .done(_ => {
+                    swalWithBootstrapButtons.fire(
+                        'Deleted!',
+                        `Todo ${todo} has been deleted`,
+                        'success'
+                    )
+                    fetchTodo()
+                })
+                .fail(xhr => {
+                    swalWithBootstrapButtons.fire(
+                        'Ops',
+                        'Todo Not found :)',
+                        'error'
+                    )
+                    console.log(xhr)
+                })
+          
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            'Cancelled',
+            'Your Todo is safe :)',
+            'error'
+          )
         }
-    })
-        .done(_ => {
-            fetchTodo()
-        })
-        .fail(xhr => {
-            console.log(xhr)
-        })
+      })
 }
